Move the casts helper file to the utils package
